added createoriginal to stage1, changed orientation_idx in displayPiece from size_t to int

+++ include/display.hpp
@@ -25,7 +25,7 @@ using Solution = std::vector<std::pair<Location, int>>; //pair<Location, orienta
 namespace dsply {
 
     //displayPiece() requires that Stage0 has already been completed (it needs the orientations to be sorted)
-    void displayPiece(const pcs::Piece* piece, size_t orientation_idx = 0);
+    void displayPiece(const pcs::Piece* piece, int orientation_idx = 0);
 
     void displayPieceList(const std::vector<std::string>& pieceList, bool allDetails = false);
